Move SimpleVector and VecArray to the cms::cuda namespace
