+++ RecoPixelVertexing/PixelTriplets/plugins/gpuFishbone.h
@@ -9,7 +9,7 @@
 
 #include "DataFormats/Math/interface/approx_atan2.h"
 #include "Geometry/TrackerGeometryBuilder/interface/phase1PixelTopology.h"
-#include "HeterogeneousCore/CUDAUtilities/interface/GPUVecArray.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/VecArray.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/cuda_assert.h"
 
 #include "GPUCACell.h"
